renamed state_space_restriction.* to likelihood_cmhn.*
