+++ doc/training/likelihood_cmhn.rst
@@ -1,20 +1,20 @@
 :orphan:
 
-The *state_space_restriction* submodule
----------------------------------------
+The *likelihood_cmhn* submodule
+-------------------------------
 
-.. automodule:: mhn.training.state_space_restriction
+.. automodule:: mhn.training.likelihood_cmhn
    :members:
    :undoc-members:
 
-.. data:: mhn.training.state_space_restriction.CUDA_AVAILABLE
+.. data:: mhn.training.likelihood_cmhn.CUDA_AVAILABLE
 
     constant returned by :code:`cuda_available()` if CUDA functions are available
 
-.. data:: mhn.training.state_space_restriction.CUDA_NOT_AVAILABLE
+.. data:: mhn.training.likelihood_cmhn.CUDA_NOT_AVAILABLE
 
     constant returned by :code:`cuda_available()` if CUDA functions are not available
 
-.. data:: mhn.training.state_space_restriction.CUDA_NOT_FUNCTIONAL
+.. data:: mhn.training.likelihood_cmhn.CUDA_NOT_FUNCTIONAL
 
     constant returned by :code:`cuda_available()` if CUDA functions don't work as expected, in this case check your CUDA drivers
