+++ mhn/__init__.py
@@ -3,8 +3,8 @@ from . import model
 from . import training
 from . import full_state_space
 
-from .training.state_space_restriction import cuda_available
-from .training.state_space_restriction import CUDA_AVAILABLE, CUDA_NOT_AVAILABLE, CUDA_NOT_FUNCTIONAL
+from .training.likelihood_cmhn import cuda_available
+from .training.likelihood_cmhn import CUDA_AVAILABLE, CUDA_NOT_AVAILABLE, CUDA_NOT_FUNCTIONAL
 
 
 def _get_metadata():
