+++ mhn/training/__init__.py
@@ -4,5 +4,5 @@ compute scores and their gradients that are needed for learning MHNs much faster
 """
 
 from . import state_containers
-from . import state_space_restriction
+from . import likelihood_cmhn
 from . import likelihood_omhn
