+++ mhn/training/likelihood_cmhn.pxd
@@ -1,6 +1,6 @@
 # by Stefan Vocht
 #
-# this is the header file for state_space_restriction.pyx
+# this is the header file for likelihood_cmhn.pyx
 
 from mhn.training.state_containers cimport State
