+++ test/test_full_state_space.py
@@ -6,7 +6,8 @@ This file contains unit tests for the full state-space functions of the mhn pack
 import unittest
 import numpy as np
 
-from mhn.training import state_space_restriction
+import mhn
+from mhn.training import likelihood_cmhn
 from mhn.full_state_space import Likelihood, ModelConstruction, PerformanceCriticalCode
 
 
@@ -19,7 +20,7 @@ class TestCudaGradient(unittest.TestCase):
         Preparation for each test
         """
         np.random.seed(0)  # set random seed for reproducibility
-        if state_space_restriction.cuda_available() != state_space_restriction.CUDA_AVAILABLE:
+        if mhn.cuda_available() != mhn.CUDA_AVAILABLE:
             self.skipTest("CUDA not available for testing")
 
     def test_compare_with_cython(self):
